ANN & CNN Models with the dataset
The ANN model was trained with the data obtained from each customer by the bank to try to make a prediction whether or not the customer would continue to bank.

In the CNN model he was trained with about 8000 images of cats and dogs. Thus having a great performance in identifying cats and dogs in images.